Created NoSEALKitPopup class in popups.py
This popup will appear when there is no SEAL kit detected.  The popup will instruct the user to make sure the SEAL kit is connected properly and provide a refresh button to allow the user to check if the SEAL is detected.  Once the SEAL kit is detected they will be permitted to move onto the next portion of the application.