Reworked how the popup windows work.
The popup windows are no longer "unclosable" — meaning there was no way to get rid of them other than completing the given prompt in the popup.  Instead the application will not be displayed until the popup windows are successfully completed.  If the user closes the popup the application will be exited.  Before the only way for users to exit the main application was for the user to close the main application window (which would open in conjunction with the popup), however, in addition to being able to close the main (which would then cause the popup to close) the user would also be able to interact with the entire main window, despite the popup window not being completed.  This was extremely undesirable behavior as the prompts in the popup window needed to be completed in order for the main app to work.  This is what prompted this change.