Fixed #1
Apparently there is a bug in the legact SEAL kit which would allow the user to apply voltage outside the acceptable voltage range (which is between 0 and 5 volts).  This commit ensures that the voltage applied can never exceed 5 volts and drop below 0 volts.  As far as I know these limits appear to be hard-coded into SEAL kit itself and I have not found a way to apply voltage beyond these limits.