Changed how LEDs are flashed
Instead of flashing them on once and only at the end of the flashing turning them of, we instead of activate cycles of LED flashing.  The data currently collected regarding LED flashes is based on interrupted cycles rather than continuous uninterrupted flashing of multiple cycles.  This change is made in order to match the data more closely.