Converted app to use PySide6 instead of PyQt6
This was an effort to be able to use the "Universal" style offered by Qt.  However, after much reseach it appears that the "Universal" style offered by Qt is only avaliable for QML based apps.  The current design of this app is a widgets based app.  There may be some type of work around which why this branch exists.  Additionally, this branch should be considered unstable as it has not been thoroughly tested.  Lastly, I have not determined how to compile the app using PySide6.